+++ bcn_rainfall_webapp/db/__init__.py
@@ -1,0 +1,3 @@
+from bcn_rainfall_webapp.db.client import DBClient
+
+__all__ = ["DBClient"]