+++ bcn_rainfall_webapp/views/__init__.py
@@ -1,3 +1,3 @@
-from .navbar import navbar
+from .sections import sections
 
-__all__ = ["navbar"]
+__all__ = ["sections"]